@@ -1,5 +1,6 @@
+import "./helpers/Prototypes";
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./components/App";
+import { App } from "./components/App.js";
 
 ReactDOM.render(<App />, document.getElementById("root"));
